#84 Add Button and Confirmation
The button is at the bottom of the account editing page. It opens a confirm modal - containing a warning and explanation - where you need to enter a given text in a text field and then click confirm again.
If the entered text is correct (case insensitive), the deletion will be started (when implemented)

CSS: Added a left margin of 180px to lign up with the other editing options